feat: modified character endpoint in a way that the user can perform crud on only Characters that he have created

+++ Role Playing Game API/Service/CharacterService.cs
@@ -15,14 +15,14 @@ namespace Role_Playing_Game_API.Service
 
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        private readonly HttpContextAccessor _httpContextAccessor;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         private static List<Character> characters = new List<Character>
         {
             new Character(),
             new Character { Id = 1, Name = "Sam" }
         };
 
-        public CharacterService(IMapper mapper, DataContext context, HttpContextAccessor httpContextAccessor)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
@@ -30,24 +30,28 @@ namespace Role_Playing_Game_API.Service
         }
 
         private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
-            .FindFirstValue(ClaimTypes.NameIdentifier)); 
+            .FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
-            Character character = _mapper.Map<Character>(newCharacter);
+            var character = _mapper.Map<Character>(newCharacter);
+
+            //Get authenticated user Id and assign it to the User field of the Character entity
+            character.User = await _context.Users.FirstOrDefaultAsync(user => user.Id == GetUserId());
 
             _context.Characters.Add(character); // Tracking data to be added to DB
             await _context.SaveChangesAsync(); // Insert changes to the DB
 
             response.Data = await _context.Characters
+                .Where(character => character.User.Id == GetUserId())
                 .Select(character => _mapper.Map<GetCharacterDto>(character))
                 .ToListAsync();
 
             return response;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters() 
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             //Only get the Character created by the authenticated user
             var response = new ServiceResponse<List<GetCharacterDto>>();
@@ -65,7 +69,8 @@ namespace Role_Playing_Game_API.Service
             var response = new ServiceResponse<GetCharacterDto>();
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(character => character.Id == id);
+                var dbCharacter = await _context.Characters
+                    .FirstOrDefaultAsync(character => character.Id == id && character.User.Id == GetUserId()); // Only get or see the character created by user
                 response.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
 
             }
@@ -81,20 +86,34 @@ namespace Role_Playing_Game_API.Service
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updateCharacter)
         {
             var response = new ServiceResponse<GetCharacterDto>();
+
             try
             {
                 var character = await _context.Characters
+                    .Include(character => character.User) // To access related entities we need to include them first
                     .FirstOrDefaultAsync(character => character.Id == updateCharacter.Id);
 
-                character.Name = updateCharacter.Name;
-                character.HitPoints = updateCharacter.HitPoints;
-                character.Strength = updateCharacter.Strength;
-                character.Defense = updateCharacter.Defense;
-                character.Class = updateCharacter.Class;
+                if (character.User.Id == GetUserId())
+                {
 
-                await _context.SaveChangesAsync();
-                 
-                response.Data = _mapper.Map<GetCharacterDto>(character);
+                    character.Name = updateCharacter.Name;
+                    character.HitPoints = updateCharacter.HitPoints;
+                    character.Strength = updateCharacter.Strength;
+                    character.Defense = updateCharacter.Defense;
+                    character.Class = updateCharacter.Class;
+
+                    await _context.SaveChangesAsync();
+
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+
+                    return response;
+                }
+
 
 
             }
@@ -114,16 +133,30 @@ namespace Role_Playing_Game_API.Service
             var response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                // FirstOrDefault return null if no elements was found while First throws an exception
+                // (FirstOrDefault) return null if no elements was found while (First) does not return null but throws an exception
                 var character = await _context.Characters
-                    .FirstAsync(character => character.Id == id);
+                    .FirstOrDefaultAsync(character => character.Id == id && character.User.Id == GetUserId());
 
-                _context.Characters.Remove(character);
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
 
-                await _context.SaveChangesAsync();
+                    response.Data = await _context.Characters
+                        .Where(character => character.User.Id == GetUserId())
+                        .Select(character => _mapper.Map<GetCharacterDto>(character)).ToListAsync();
+                }
+                else
+                {
 
-                response.Data = await _context.Characters
-                    .Select(character => _mapper.Map<GetCharacterDto>(character)).ToListAsync();
+                    response.Success = false;
+                    response.Message = "Character not found";
+
+                    return response;
+
+                }
+
+
 
             }
             catch (Exception ex)
